RaspBerryPiVersion/version1/requirements.txt
numpy~=1.26.4
zivid~=2.11.1.2.11.1
fanucpy~=0.1.14
pillow~=10.2.0
CTkToolTip~=0.8
customtkinter~=5.2.2
CTkMenuBar~=0.8
matplotlib~=3.8.3
PyYAML~=6.0.1